+++ packages/expo-modules-core/ios/Core/Views/SwiftUI/SwiftUIHostingView.swift
@@ -85,6 +85,7 @@ extension ExpoSwiftUI {
       }
 
       props.shadowNodeProxy = shadowNodeProxy
+      props.hostingView = self
 
       shadowNodeProxy.objectWillChange.send()